requirements.txt
Django==5.0.7
djangorestframework
requests
psycopg2-binary
gunicorn
whitenoise
boto3==1.34.131
django-storages==1.14.2
python-dotenv==1.0.0
Pillow==10.0.0
whitenoise==6.5.0